index.htm from BlunderDome/wingnut for testing
This was created by a page save as .htm, and Edge rewrote some of it to include the linked images in a folder.  This will not be the case when extracting from the FrontPage/IIS/VSS of the pages.  I will need to deal with that.  For the present case, I simply want to know that index.htm can be accessed.